chore(cleanup): Split utils index into multiple files
